refactor: rename group -> team
