+++ src/main/java/CloudProject/A_meet/domain/group/domain/team/domain/UserTeam.java
@@ -1,34 +1,31 @@
-package CloudProject.A_meet.domain.group.domain.group.domain;
+package CloudProject.A_meet.domain.group.domain.team.domain;
 
+import CloudProject.A_meet.domain.group.domain.user.domain.User;
 import CloudProject.A_meet.global.common.model.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
-import org.springframework.cglib.core.Local;
-import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import java.time.LocalDateTime;
-
-@Table(name="teams")
+@Table(name="user_team")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 @ToString
 @Entity
 @EntityListeners(AuditingEntityListener.class)
-public class Team extends BaseTimeEntity {
+public class UserTeam extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(nullable = false)
-    private Long teamId;
+    private Long userTeamId;
 
-    @Column(nullable = false)
-    private String name;
+    @ManyToOne
+    @JoinColumn(name="userId")
+    private User userId;
 
-    private String description;
-
-    @Column(nullable = false)
-    private long maxPeople;
+    @ManyToOne
+    @JoinColumn(name="team_id")
+    private Team teamId;
 }
